+++ Electrode/src/Electrode_CSTR.cpp
@@ -917,7 +917,7 @@ void Electrode_CSTR::updateSpeciesMoleChangeFinal()
     std::fill(DspMoles_final_.begin(), DspMoles_final_.end(), 0.0);
     double mult = (surfaceAreaRS_init_[0] + surfaceAreaRS_final_[0]);
     mult /= 2.0;
-    for (int i = 0; i < m_totNumVolSpecies; i++) {
+    for (size_t i = 0; i < m_totNumVolSpecies; i++) {
         DspMoles_final_[i] += mult * spNetProdPerArea[i];
     }
